Isuue 14 y 15 sprint 1 ( base de datos y endpoint de registro)

+++ src/api/routes.py
@@ -2,8 +2,11 @@
 This module takes care of starting the API Server, Loading the DB and Adding the endpoints
 """
 from flask import Flask, request, jsonify, url_for, Blueprint
+
+from api.models import db, User, Role, Project, UserProjectRole, TokenBlockedList
+=======
 from datetime import timedelta
-from api.models import db, User, Role, Project, TokenBlockedList
+
 from api.utils import generate_sitemap, APIException
 from flask_cors import CORS
 from flask_jwt_extended import create_access_token, jwt_required, get_jwt, get_jwt_identity
@@ -28,46 +31,37 @@ mail = Mail(app)
 # Allow CORS requests to this API
 CORS(api)
 
-@api.route("/user", methods=["POST"])
-def user_create():
-    try:
-        body = request.get_json()
-
-        required_fields = ["full_name", "email", "password"]
-        for field in required_fields:
-            if field not in body or body[field] is None:
-                return jsonify({"msg": f"Debe especificar un {field}"}), 400
-
-        user = User.query.filter_by(email=body["email"]).first()
-        if user is not None:
-            return jsonify({"msg": "Usuario ya existe"}), 400
-
-        body["password"] = bcrypt.generate_password_hash(body["password"]).decode("utf-8")
-        user = User(full_name=body["full_name"], email=body["email"], password=body["password"], is_active=True)
-        db.session.add(user)
-        db.session.commit()
-
-        return jsonify({"msg": "Usuario creado", "user": user.serialize()}), 200
-    except Exception as e:
-        return jsonify({"msg": str(e)}), 500
 
 @api.route("/user", methods=["PATCH"])
 def user_modified():
     body= request.get_json()
 
-    required_fields = ["email", "usuario", "password", "role"]
+    required_fields = ["email", "full_name", "password", "role"]
     for field in required_fields:
         if field not in body:
             return jsonify({"msg": f"Falta el campo: {field}"}), 400
         
     user = User.query.filter_by(email=body["email"]).first()
-    if user is not None:
-        user.password = bcrypt.generate_password_hash(body["password"]).decode("utf-8")
-        user.usuario = body["usuario"]
-        user.role = body["role"]
-        db.session.commit()
-        return jsonify({"msg": "Usuario modificado", "user": user.serialize()})
-    return jsonify({"msg": "Usuario no encontrado"}), 404
+    if user is None:
+        return jsonify({"msg": "Usuario no encontrado"}), 404
+        
+    user.full_name = body["full_name"]
+    user.password = bcrypt.generate_password_hash(body["password"]).decode("utf-8")
+    user.email = body["email"]
+       
+    role_to_assign = Role.query.filter_by(name=body["role"]).first()
+    if not role_to_assign:
+        return jsonify({"msg": "Rol no encontrado"}), 404
+    
+    # Evitar que se asigne el rol de Administrador de Plataforma
+    if role_to_assign.name == "Administrador de Plataforma":
+        return jsonify({"msg": "No se puede asignar el rol de Administrador de Plataforma"}), 403
+
+    user.global_role = role_to_assign
+        
+    db.session.commit()
+    return jsonify({"msg": "Usuario modificado", "user": user.serialize()}),200
+    
 
 @api.route("/user", methods=["DELETE"])
 def user_delete():
@@ -82,6 +76,10 @@ def user_delete():
         if user is None:
             return jsonify({"msg": "Usuario no existe"}), 404
 
+        # Verificar que el usuario no sea un Administrador de Plataforma
+        if user.global_role and user.global_role.name == "Administrador de Plataforma":
+            return jsonify({"msg": "No se puede eliminar un Administrador de Plataforma"}), 403
+        
         db.session.delete(user)
         db.session.commit()
 
@@ -182,57 +180,62 @@ def project_create():
 
 
 
-@api.route("/register/initial-admin", methods=["POST"])
-def register_initial_admin():
+@api.route("/register", methods=["POST"])
+def register_user():
     body = request.get_json()
+
     
-    # Verificar campos requeridos.
-    required_fields = ["full_name", "email", "password", "project_name"]
+
+    required_fields = ["email", "password", "full_name"]
+
     for field in required_fields:
-        if field not in body or body[field] is None:
-            return jsonify({"msg": f"Debe especificar un {field}"}), 400
+        if field not in body or not body[field]:
+            return jsonify({"msg": f"Falta el campo: {field}"}), 400
+
     
-    # Verificar si el usuario ya existe
     existing_user = User.query.filter_by(email=body["email"]).first()
     if existing_user:
         return jsonify({"msg": "El usuario ya existe"}), 400
 
-    # Verificar si el proyecto ya existe
-    existing_project = Project.query.filter_by(name=body["project_name"]).first()
-    if existing_project:
-        return jsonify({"msg": "El proyecto ya existe"}), 400
     
-    body["password"] = bcrypt.generate_password_hash(body["password"]).decode("utf-8")
+    hashed_password = bcrypt.generate_password_hash(body["password"]).decode("utf-8")
+
     
-    # Crear el rol de administrador si no existe
-    admin_role = Role.query.filter_by(name='admin').first()
-    if not admin_role:
-        admin_role = Role(name='admin', description='Administrador con acceso completo')
-        db.session.add(admin_role)
-        db.session.commit()
-    
-    # Crear el nuevo usuario administrador
-    new_admin = User(
-        full_name=body["full_name"],
+
+    new_user = User(
+
         email=body["email"],
-        password=body["password"],
-        is_active=True
+        password=hashed_password,
+        full_name=body["full_name"],
+        is_active=True,
+        global_role=None  # No se asigna un rol global
     )
-    new_admin.roles.append(admin_role)
+
+    # Si es el primer usuario, asignarle rol de Administrador de Plataforma
+    if User.query.count() == 0:
+
+        # Verificar si el rol "Administrador de Plataforma" existe en la tabla `Role`
+        admin_role = Role.query.filter_by(name='Administrador de Plataforma').first()
+        if not admin_role:
+            # Crear el rol si no existe
+            admin_role = Role(name='Administrador de Plataforma', description='Acceso completo a la plataforma')
+            db.session.add(admin_role)
+            db.session.commit()
+
+        # Asignar el rol "Administrador de Plataforma" al primer usuario
+        new_user.global_role = admin_role
     
-    # Crear el proyecto asociado al administrador
-    new_project = Project(
-        name=body["project_name"],
-        description=body.get("project_description", ""),
-        admin=new_admin
-    )
-    
-    # Guardar los cambios en la base de datos
-    db.session.add(new_admin)
-    db.session.add(new_project)
+    db.session.add(new_user)
     db.session.commit()
+
+
+    return jsonify({
+        "msg": "Usuario registrado exitosamente",
+        "user": new_user.serialize()
+    }), 200
+
     
-    return jsonify({"msg": "Administrador y proyecto creados", "user": new_admin.serialize(), "project": {"name": new_project.name, "description": new_project.description}}), 200
+   
 
 @api.route("/login", methods=['POST'])
 def user_login():
